feat: add tptctl replace command, remove describe command (#371)
* dev: move magefiles to magefiles dir

By moving the magefiles to their own directory, we can remove the build
tags and elliminate the erroneous build warnings while preserving the
convenient mage functionality.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* refactor: move Threeport config code to pkg/cli/v0 package

The package `pkg/config/v0` contains the configuration abstractions for
CLI objects to make interacting with those objects more human-friendly.
The package `pkg/cli/v0` contains tptctl functionality.
The Threeport config lives on a user's filesystem and is used by tptctl
to connect to the Threeport API. It is far more sensible for this to
live in the cli package.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* feat: add Replace function to client lib for all objects

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* refactor: return config objects from all config abstraction methods

The config abstraction methods previously returned API objects which the
client (cmd/tptctl/cmd package) had to then translate into user-relevant
info.  This also did not give a user a way to get configs they could
modify and then submit as a part of a change.  This issue is remedied
with this commit.
This commit adds Get and Replace methods for each config
abstraction.
This commit removes the Describe methods.  They are no longer
needed.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* refactor: improve code reuse and brevity for instance config abstractions

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* refactor: return defined instance config abstraction from config pkg

Previously, the defined instance config abstraction CRUD methods
returned the both the definition and instance objects. Now, they return
the defined instance config abstraction, pulling all the logic of how to
map definition and instance attributes into the defined instance fields.
This will simplify the CLI code - and anything else that uses the config
package, keep the config abstraction logic in one place.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* feat: config package to produce useful yaml and json object output

This commit adds the config package *Values object field tags to omit
nil value fields for YAML and JSON. . It also adds the ability to handle
encrypted fields for output with the `get` commands.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* refactor: remove describe cmd, return config objects from get cmd

This commit overhauls the tptctl CLI so that config objects from the
config package are returned to the user from the `get` command, rather
than select fields from the API objects.  This way, the user can
retrieve objects using tptctl and make modifications to them in order
to make changes since the input required is a config object.

This significantly cleans up the contract between tptctl and config
package.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* chore: remove backup files from config changes

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* fix: implement secret decryption for terraform def inst

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* fix: add missing field to instance values objects in config pkg

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* refactor: use Config object for config package method receivers

Now each method for get, create, replace and delete objects in the
config package uses the Config object as the receiver.  Previously, it
had the Values object as the receiver while returning the Config
package.  This change was made for improved consistency and greater ease
of use by clients of this package.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* fix: repair SDK's CLI compatibility with config package

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* test: fix panic in integration test

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

---------

Signed-off-by: Rich Lander <lander2k2@protonmail.com>