chore: remove VERSION column from tptctl tabular output (#382)
* chore: remove VERSION column from tptctl tabular output

The version column was intended to inform the user which version of an
API object was being displayed.  It was an attempt to future-proof for
when multiple versions exist.  However, it was just confusing.  We will
tackle how to distinguish different API object versions when there are
more versions to manage.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* fix: remove potential nil pointer errors on tptctl ouput

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

---------

Signed-off-by: Rich Lander <lander2k2@protonmail.com>